pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ca2roi"
version = "0.1.0"
description = "Calcium imaging ROI and bleaching analysis CLI tool"
authors = []
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.2.1",
    "matplotlib>=3.10.3",
    "numpy>=2.3.1",
    "opencv-python>=4.11.0.86",
    "pandas>=2.0.0",
    "pillow>=11.3.0",
    "scikit-image>=0.25.2",
    "scikit-learn>=1.3.0",
    "scipy>=1.11.0",
]

[project.scripts]
ca2roi = "ca2roi.cli:main"


[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=src/ca2roi --cov-report=term-missing --cov-report=html"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by formatter
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
